fromSource: prevent errors with large source maps
Using a regex to match against very long files can cause the process to
run out of memory with a "call stack size exceeded" error.
This patch adds a `largeSource` flag that switches from a regex match to
a `indexOf` match.

It's slightly less accurate (somewhat prone to false-positives),
but practically this shouldn't cause many problems:
  1) the source map we care about is always at the end of the file
  2) it's unlikely enough this precise string would appear in any code.

- add `largeSource` flag for `fromSource` to the docs