Make the example more sensible
This refers to the example in example/comment-to-json.js, which is also
copied into README.md.

Previously, "sources" and "sourcesContent" were mixed up. ("sources" is
an array of source file URLs, while "sourcesContent" is an optional
array containing the contents of the files listed in "sources".)

This also switches away from the deprecated `//@` syntax to `//#`, and
uses more realistic file names in the source map examples.